+++ packages/dialtone-vue2/directives/tooltip/tooltip.js
@@ -1,4 +1,4 @@
-import { DtTooltip, TOOLTIP_DIRECTIONS } from '@/components/tooltip';
+import { DtTooltip } from '@/components/tooltip';
 import { getUniqueString } from '@/common/utils';
 
 export const DtTooltipDirective = {
@@ -18,8 +18,16 @@ export const DtTooltipDirective = {
       },
 
       methods: {
-        addTooltip (id, message, placement) {
-          this.tooltips.push({ id, message, placement });
+        addOrUpdateTooltip (id, message, placement) {
+          const index = this.tooltips.findIndex(tooltip => tooltip.id === id);
+          if (index !== -1) {
+            // Update existing tooltip
+            this.tooltips[index].message = message;
+            this.tooltips[index].placement = placement;
+          } else {
+            // Add new tooltip
+            this.tooltips.push({ id, message, placement });
+          }
         },
 
         removeTooltip (id) {
@@ -54,42 +62,30 @@ export const DtTooltipDirective = {
 
     DtTooltipDirectiveApp.$mount(mountPoint);
 
-    const isValidBindingTextValue = (value) => typeof value === 'string' && value?.trim();
-    const isValidBindingPlacementValue = (value) => value === undefined || TOOLTIP_DIRECTIONS.includes(value);
-
     Vue.directive('dt-tooltip', {
       bind (anchor, binding) {
-        if (!isValidBindingTextValue(binding.value)) {
-          // eslint-disable-next-line no-console
-          console.warn(
-            'Missing value for v-dt-tooltip directive on: ',
-            anchor,
-            'received value: ',
-            binding.value,
-          );
-          return;
+        // Initial tooltip setup
+        setupTooltip(anchor, binding);
+      },
+      update (anchor, binding) {
+        // Update tooltip on binding value change
+        if (binding.value !== binding.oldValue) {
+          setupTooltip(anchor, binding);
         }
-        if (!isValidBindingPlacementValue(binding.arg)) {
-          // eslint-disable-next-line no-console
-          console.warn(
-            'Wrong placement value provided for v-dt-tooltip directive on: '
-            , anchor,
-            'received value: ',
-            binding.arg);
-          return;
-        }
-
-        const tooltipId = getUniqueString();
-        const message = binding.value;
-        const placement = binding.arg || DEFAULT_PLACEMENT;
-
-        anchor.setAttribute('data-dt-tooltip-id', tooltipId);
-        DtTooltipDirectiveApp.addTooltip(tooltipId, message, placement);
       },
       unbind (anchor) {
         DtTooltipDirectiveApp.removeTooltip(anchor.getAttribute('data-dt-tooltip-id'));
       },
     });
+
+    function setupTooltip (anchor, binding) {
+      const tooltipId = anchor.getAttribute('data-dt-tooltip-id') || getUniqueString();
+      const message = binding.value;
+      const placement = binding.arg || DEFAULT_PLACEMENT;
+
+      anchor.setAttribute('data-dt-tooltip-id', tooltipId);
+      DtTooltipDirectiveApp.addOrUpdateTooltip(tooltipId, message, placement);
+    }
   },
 };